requirements.txt
torch>=1.9.0
numpy>=1.21.0
opencv-python>=4.5.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.0.0
pyyaml>=5.4.0
matplotlib>=3.3.0
Pillow>=8.0.0
tqdm>=4.62.0
scikit-learn>=0.24.0  # если нужны метрики в будущем